# web-v2/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiEnvelope_FileEventFacetsData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileEventFacetsData"
              },
              {
                "type": "null"
              }
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "data",
          "errors"
        ],
        "title": "ApiEnvelope[FileEventFacetsData]",
        "type": "object"
      },
      "ApiEnvelope_FileEventGroupsData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileEventGroupsData"
              },
              {
                "type": "null"
              }
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "data",
          "errors"
        ],
        "title": "ApiEnvelope[FileEventGroupsData]",
        "type": "object"
      },
...
      "ApiEnvelope_PaginatedData_FileEventResource__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_FileEventResource_"
              },
              {
                "type": "null"
              }
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "data",
          "errors"
        ],
        "title": "ApiEnvelope[PaginatedData[FileEventResource]]",
        "type": "object"
      },
...
      "ApiEnvelope_PaginatedData_RunHeader__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_RunHeader_"
              },
              {
                "type": "null"
              }
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "data",
          "errors"
        ],
        "title": "ApiEnvelope[PaginatedData[RunHeader]]",
        "type": "object"
      },
...
      "ApiEnvelope_RunFacetsData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunFacetsData"
              },
              {
                "type": "null"
              }
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "data",
          "errors"
        ],
        "title": "ApiEnvelope[RunFacetsData]",
        "type": "object"
      },
...
      "ApiError": {
        "additionalProperties": false,
        "description": "One stable Web API error or disabled reason.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "remediation": {
            "$ref": "#/components/schemas/ApiRemediation",
            "title": "Remediation"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ApiFailureEnvelope": {
        "additionalProperties": false,
        "description": "Failure-only envelope used for declared error responses.",
        "properties": {
          "data": {
            "title": "Data",
            "type": "null"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "minItems": 1,
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "data",
          "errors"
        ],
        "title": "ApiFailureEnvelope",
        "type": "object"
      },
      "ApiRemediation": {
        "additionalProperties": false,
        "description": "Optional recovery action displayed but never executed automatically.",
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "route": {
            "title": "Route",
            "type": "string"
          }
        },
        "required": [
          "label"
        ],
        "title": "ApiRemediation",
        "type": "object"
      },
...
      "FileEventFacetsData": {
        "additionalProperties": false,
        "description": "FileEvent status facets plus total events for one Run.",
        "properties": {
          "facets": {
            "$ref": "#/components/schemas/FileEventFacetSets"
          },
          "total": {
            "$ref": "#/components/schemas/NonNegativeCount"
          }
        },
        "required": [
          "facets",
          "total"
        ],
        "title": "FileEventFacetsData",
        "type": "object"
      },
...
      "FileEventStatus": {
        "description": "Supported durable operation event statuses.",
        "enum": [
          "pending",
          "succeeded",
          "failed"
        ],
        "title": "FileEventStatus",
        "type": "string"
      },
...
      "PaginatedData_FileEventResource_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FileEventResource"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "PaginatedData[FileEventResource]",
        "type": "object"
      },
...
      "PaginatedData_RunHeader_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunHeader"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "PaginatedData[RunHeader]",
        "type": "object"
      },
...
      "RunFacetsData": {
        "additionalProperties": false,
        "description": "Run status facets plus total Runs in scope.",
        "properties": {
          "facets": {
            "$ref": "#/components/schemas/RunFacetSets"
          },
          "total": {
            "$ref": "#/components/schemas/NonNegativeCount"
          }
        },
        "required": [
          "facets",
          "total"
        ],
        "title": "RunFacetsData",
        "type": "object"
      },
...
      "RunStatus": {
        "description": "Supported Run statuses.",
        "enum": [
          "running",
          "succeeded",
          "partial_failed",
          "failed"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "OMYM2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/history": {
      "get": {
        "operationId": "getHistory",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "plan_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plan Id"
            }
          },
          {
            "in": "query",
            "name": "library_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Library Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_PaginatedData_RunHeader__"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get History"
      }
    },
    "/api/history/facets": {
      "get": {
        "operationId": "getHistoryFacets",
        "parameters": [
          {
            "in": "query",
            "name": "library_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Library Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_RunFacetsData_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get History Facets"
      }
    },
...
    "/api/history/{run_id}/events": {
      "get": {
        "operationId": "getRunEvents",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FileEventStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_PaginatedData_FileEventResource__"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Run Events"
      }
    },
    "/api/history/{run_id}/events/facets": {
      "get": {
        "operationId": "getRunEventFacets",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_FileEventFacetsData_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Run Event Facets"
      }
    },
    "/api/history/{run_id}/events/groups": {
      "get": {
        "operationId": "getRunEventGroups",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "group_by",
            "required": true,
            "schema": {
              "const": "target_directory",
              "title": "Group By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_FileEventGroupsData_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Run Event Groups"
      }
    },
...
  }
}